updated jet ID

+++ AllHadronicSUSY/Utils/python/goodjetsproducer_cfi.py
@@ -2,7 +2,7 @@ import FWCore.ParameterSet.Config as cms
 
 GoodJetsProducer = cms.EDProducer('GoodJetsProducer',
 JetTag               = cms.InputTag('slimmedJets'),
-maxMuFraction								  = cms.double(30),
+maxMuFraction								  = cms.double(2),
 minNConstituents								  = cms.double(1),
 maxNeutralFraction								  = cms.double(0.99),
 maxPhotonFraction								  = cms.double(0.99),
